requirements.txt
matplotlib==3.7.1
natsort==8.3.1
numpy==1.24.2
onnxruntime==1.14.1
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.2
tensorflow==2.11.0
timm==0.6.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
